A few interface changes for how to construct an EOF error, and a system error based on an errno value.
Before, the constructor was grabbing the system errno value.  Well, it
turns out there were often system calls being made before it got to that
part of the constructor, and these calls would wipe out the value of
errno.

So, I changed it so there's a static function of the UNIXError class for
grabbing errno values (so it's easy to change how it happens in case
threading or some other change to the standard library mucks things up a
lot), and changed the constructor to require an errno value.  This
arrange it so there was no longer any overlap between the EOF error
constructor and the Unix system error constructor.